host/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "dacdemo"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "pyserial>=3.5",
    "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
dacdemo = "dacdemo.cli:main"

[tool.setuptools.packages.find]
where = ["."]
